set viewport to ping location in single ping page
